refacor 1
