+++ anifetch-cli
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+
+'''
+    Temporary CLI entry point for anifetch.
+'''
+
+from anifetch import main  # ou core.main selon ta structure
+
+if __name__ == "__main__":
+    main()
