+++ anifetch/__init__.py
@@ -1,0 +1,12 @@
+# anifetch/__init__.py
+
+'''
+    Anifetch package initialization module.
+'''
+
+from .core import run_anifetch
+from .cli import parse_args
+
+def main():
+    args = parse_args()
+    run_anifetch(args)
