+++ anifetch/__main__.py
@@ -1,0 +1,11 @@
+# anifetch/main.py
+
+'''
+    Anifetch main module for running the application.
+'''
+
+
+from . import main
+
+if __name__ == "__main__":
+    main()
