+++ scripts/anifetch-static-resize2.sh
